Experiment: Interruptible AHK commands
This essentially reverts the following commits:

- 55a95e0
- b4a2ae5
- a6b013e
- a478d4f

The trick to interruptible AHK commands is setting the timer to check
for signals *before* calling the main Python script. This makes sense,
because the only time the AHK timers tick during the main Python code,
is when the Python code calls a blocking AHK command.

However, the biggest downside of this approach is that the signal that
was detected by the timer cannot interrupt the blocking AHK command that
allowed this timer to execute. Thus, there's no way to raise
KeyboardInterrupt during the AHKCall.