Fix module search paths in conda
When AutoHotkey.py is started in a conda environment, it fails silently
because it cannot import the encodings module. Checking the redirected
output, I noticed that during initialization sys.path contains the
following items: '.\DLLs' and '.\lib'. I think that the embedded
interpreter didn't infer the Python installation directory. Setting the
path explicitly in the PYTHONHOME environment variable fixes the
problem.